Finalizing Register Screen UI and moved TextField class to component package.
