+++ app/src/main/java/id/arvigo/arvigobasecore/di/AppModules.kt
@@ -1,10 +1,12 @@
 package id.arvigo.arvigobasecore.di
 
 import id.arvigo.arvigobasecore.data.source.network.ApiService
+import id.arvigo.arvigobasecore.ui.feature.login.LoginViewModel
 import id.arvigo.arvigobasecore.util.Constant.BASE_URL
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import org.koin.androidx.compose.BuildConfig
+import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
@@ -30,3 +32,6 @@ val networkModule = module {
     }
 }
 
+val viewModelModules = module {
+    viewModel { LoginViewModel() }
+}
